remove sendhumanpose files
