--- demos/human_pose_estimation_demo/include/send_human_pose.hpp
@@ -1,15 +1,0 @@
-// Copyright (C) 2018-2019 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-#pragma once
-
-#include <vector>
-
-#include <opencv2/core/core.hpp>
-
-#include "human_pose.hpp"
-
-namespace human_pose_estimation {
-    void sendHumanPose(const std::vector<HumanPose>& poses);
-}  // namespace human_pose_estimation
